REFACTOR: Moved models list constant into `constants` folder
