+++ components/chat/message-bubble.tsx
@@ -6,7 +6,7 @@ import Link from "next/link";
 import { useAppStore } from "@/lib/store";
 import type { Message } from "@/types/message";
 import { ROUTES } from "@/lib/routes";
-import { MODELS } from "@/models";
+import { MODELS } from "@/constants/models";
 import { Bot, Command, User } from "lucide-react";
 import { useMemo, useState } from "react";
 import { MarkdownRenderer } from "./markdown-renderer";
